Fixes #1986 #2029: Allow beetbox base, fix drush aliases in Vm." (#2033)
git commit -m Fixes
